440: Split users controller into separate files for each use case
