+++ apps/backend/src/routes/users.routes.ts
@@ -1,11 +1,9 @@
 import { Router } from 'express';
-import {
-  createUser,
-  deleteUser,
-  login,
-  logout,
-  refreshTokens,
-} from '../controllers/users.controllers.js';
+import { createUser } from '../controllers/users/createUser.uc.js';
+import { deleteUser } from '../controllers/users/deleteUser.uc.js';
+import { login } from '../controllers/users/login.uc.js';
+import { logout } from '../controllers/users/logout.uc.js';
+import { refreshTokens } from '../controllers/users/refreshTokens.uc.js';
 import { acceptBodyCheck } from '../middlewares/acceptBodyCheck.js';
 import { acceptHeaderCheck } from '../middlewares/acceptHeaderCheck.js';
 import { authenticateAccessToken } from '../middlewares/auth.js';
